supervisor: a simple health check and restart service
Any time crazytree receives a packet from the server, it sends a
"heartbeat" byte to the supervisor via stdout. If the supervisor
sees no heartbeat for 60s, crazytree is restarted in 30s.

If crazytree hangs hard (can it?) then supervisor hangs as well
as we make process-wait wait forever.

If Hurnscald is really quiet (and no user there), perhaps no
packets will be received and crazytree will be dead.